% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupPtcomp.R
\name{addCompExp}
\alias{addCompExp}
\title{addCompExp}
\usage{
addCompExp(meta.df, exp.mat, unit.num)
}
\arguments{
\item{meta.df}{tibble, data frame containing cell information}

\item{exp.mat}{numeric matrix, gene (row) * cell (column) matrix}

\item{unit.num}{numeric, how many cells are calculated for one entry of mean expression entry}
}
\value{
comp.exp.mat numeric matrix, expression for several cells
}
\description{
Add columns of compressed expression
}
\details{
Add columns of compressed expression
}
\author{
Yasuhiro Kojima
}
