% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupPtcomp.R
\name{calculateCompExpMat}
\alias{calculateCompExpMat}
\title{calculateCompExpMat}
\usage{
calculateCompExpMat(exp.mat, unit.num)
}
\arguments{
\item{exp.mat}{numeric matrix, gene (row) * cell (column) matrix}

\item{unit.num}{numeric, how many cells are calculated for one entry of mean expression entry}
}
\value{
comp.exp.mat numeric matrix, compressed expression matrix
}
\description{
Calculate compressed expression
}
\details{
Calculate compressed expression from \code{exp.mat}.
}
\author{
Yasuhiro Kojima
}
