% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupPtcomp.R
\name{calculateCompTVec}
\alias{calculateCompTVec}
\title{calculateCompTVec}
\usage{
calculateCompTVec(t.vec, unit.num)
}
\arguments{
\item{t.vec}{numeric matrix, pseudo time for each cell}

\item{unit.num}{numeric, how many cells are calculated for one entry of mean expression entry}
}
\value{
comp.t.vec numeric vecotor, compressed pseudo time
}
\description{
Calculate compressed pseudo time
}
\details{
Calculate compressed pseudo time.
}
\author{
Yasuhiro Kojima
}
