% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussianProcess.R
\name{calculateGP}
\alias{calculateGP}
\title{calculateGP}
\usage{
calculateGP(exp.vec, t.vec)
}
\arguments{
\item{exp.vec}{numeric vector, expression values}

\item{t.vec}{numeric vector, pseudo time values}
}
\value{
gp.rlt list, contain loglikelihood and optimized parameters
}
\description{
Calculate GP for each gene
}
\details{
This function calculate the log likelihood and optimized parameters as list
}
\author{
Yasuhiro Kojima
}
