% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussianProcess.R
\name{calculateGPDf}
\alias{calculateGPDf}
\title{calculateGPDf}
\usage{
calculateGPDf(ptcomp.df)
}
\arguments{
\item{ptcomp.df}{data frame, contain expression and annotation}
}
\value{
gp.df data frame
}
\description{
Calculate GP for all genes and store them in data frame
}
\details{
The columns represent gene name, log likelihood and each parameter
}
\author{
Yasuhiro Kojima
}
