% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffPt.R
\name{calculateLLDiff}
\alias{calculateLLDiff}
\title{calculateLLDiff}
\usage{
calculateLLDiff(treatment1, treatment2, gene.vec, ptcomp.df)
}
\arguments{
\item{treatment1}{character, first treatment}

\item{treatment2}{character, second treatment}

\item{gene.vec}{character vec, gene names used in this comparison}

\item{ptcomp.df}{data frame, contain expression and annotation}
}
\value{
diff.ll.df data frame, gene name, treatment1 ll, treatment2 ll, merged ll
}
\description{
Compare the LL for separated or merged model
}
\details{
Calcualte LL for each treatment and merginalized data, and see the difference
}
\author{
Yasuhiro Kojima
}
