% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussianProcess.R
\name{concatExpMat}
\alias{concatExpMat}
\title{concatExpMat}
\usage{
concatExpMat(gene.vec, ptcomp.df)
}
\arguments{
\item{gene.vec}{charactor vector, gene names}

\item{ptcomp.df}{data frame, contain expression and annotation}
}
\value{
exp.mat
}
\description{
Concatenate expression matrix included in ptcomp data frame
}
\author{
Yasuhiro Kojima
}
