% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussianProcess.R
\name{concatTVec}
\alias{concatTVec}
\title{concatTVec}
\usage{
concatTVec(pt.comp.df)
}
\arguments{
\item{pt.comp.df}{data frame, contain expression and annotation}
}
\value{
t.vec  numeric vector, each element represent pseudo time for each gene
}
\description{
Concatenate pseudo time vector included in ptcomp data frame
}
\author{
Yasuhiro Kojima
}
