% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupPtcomp.R
\name{makeMetaData}
\alias{makeMetaData}
\title{makeMetaData}
\usage{
makeMetaData(cell.vec, t.vec, treatment.vec, replicate.vec)
}
\arguments{
\item{cell.vec}{charactor vecotor, cell index for each cell}

\item{t.vec}{numeric vector, pseudo time estimation for each cell}

\item{treatment.vec}{character vector, experimental treatment for each cell}

\item{replicate.vec}{character vector, experimental replicate for each cell}
}
\value{
meta.df
}
\description{
Set up meta data
}
\details{
Set up meta data for ptcomp calculation. Basically, it group cells by replicate
}
\author{
Yasuhiro Kojima
}
