% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffPt.R
\name{permutateTreatment}
\alias{permutateTreatment}
\title{permutateTreatment}
\usage{
permutateTreatment(ptcomp.df)
}
\arguments{
\item{ptcomp.df}{data frame, contain expression and annotation}
}
\value{
permutated.ptcomp.df data frame, contain expression and annotation
}
\description{
Permutate treatment labels
}
\author{
Yasuhir Kojima
}
