% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupPtcomp.R
\name{scaleExpression}
\alias{scaleExpression}
\title{scaleExpression.}
\usage{
scaleExpression(count.mat)
}
\arguments{
\item{count.mat}{numeric matrix, gene (row) * cell (column) matrix}
}
\value{
scale.exp.mat numeric matrix, expression
}
\description{
Scaling gene expresiion
}
\details{
Scaling the expression mean to 1 for each column (cells)
}
\author{
Yasuhiro Kojima
}
