% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussianProcess.R
\name{vec2mat}
\alias{vec2mat}
\title{vec2mat}
\usage{
vec2mat(vec)
}
\arguments{
\item{vec}{numeric vector}
}
\value{
mat numeric matrix
}
\description{
Convert vector to 1 column matrix
}
\author{
小嶋泰弘
}
